Parser testé et approuvé (un peu), marche pour les 3 documents XML (petit plan, moyen plan, grand plan). Un petit test est fait en sale depuis le main. Manque l'IHM

+++ src/main/java/fr/insa/colisvif/xml/CityMapParser.java
@@ -8,7 +8,7 @@ import org.xml.sax.SAXException;
 import javax.xml.parsers.ParserConfigurationException;
 
 public interface CityMapParser {
-    List<Triplet<Integer, Double, Double>> readNodes();
-    List<Quadruplet<Double, String, Integer, Integer>> readSections();
+    List<Triplet<Long, Double, Double>> readNodes();
+    List<Quadruplet<Double, String, Long, Long>> readSections();
     void loadFile() throws ParserConfigurationException, SAXException, IOException;
 }
